+++ examples/01-differing-versions/xiden-workspace/etc/xiden.rkt
@@ -1,4 +1,6 @@
-#lang xiden/rcfile
+#lang racket/base
+
+(provide (all-defined-out))
 
 ; Because the inputs are local
 (define XIDEN_TRUST_BAD_DIGEST #t)