+++ examples/02-package-dependencies/xiden-workspace/etc/xiden.rkt
@@ -1,3 +1,5 @@
-#lang xiden/rcfile
+#lang racket/base
+
+(provide (all-defined-out))
 
 (define XIDEN_TRUST_BAD_DIGEST #t)