+++ examples/03-input-checking/xiden-workspace/etc/xiden.rkt
@@ -1,4 +1,6 @@
-#lang xiden/rcfile
+#lang racket/base
+
+(provide (all-defined-out))
 
 ; Paste the expression from your clipboard into this list.
 ; This is how you tell Xiden that you trust a public key.