+++ rc.rkt
@@ -10,6 +10,7 @@
          "file.rkt"
          "integrity.rkt"
          "openssl.rkt"
+         "plugin.rkt"
          "setting.rkt"
          "string.rkt"
          "url.rkt"
@@ -17,14 +18,11 @@
 
 (provide
  (contract-out
-  [call-with-rcfile
-   (-> (-> any) any)]
   [XIDEN_SETTINGS
    (hash/c symbol? setting? #:immutable #t)]
   [dump-xiden-settings
    (-> (hash/c symbol? any/c))]))
 
-(define current-xiden-rcfile-cache (make-parameter void))
 
 ; This gets build out as a side-effect of instantiation. If you build
 ; it declaratively, then it's easy to forget to add a new setting
@@ -49,10 +47,10 @@
            maybe-not-void))))
 
 
-; Checks environment variable, then rcfile, then hard-coded value.
+; Checks environment variable, then plugin, then hard-coded value.
 (define (in-xiden-setting-value-sources id default-value)
   (in-list (list (λ () (maybe-get-setting-value-from-envvar (symbol->string id)))
-                 (λ () (with-handlers ([values void]) ((current-xiden-rcfile-cache) id)))
+                 (λ () (plugin-ref id (void)))
                  (const default-value))))
 
 
@@ -61,26 +59,6 @@
   (cond [(not env) (void)]
         [(string=? env "") (void)]
         [else (read (open-input-string env))]))
-
-
-(define (get-xiden-settings-path)
-  (build-workspace-path "etc/xiden.rkt"))
-
-
-; Warning: This performs a full file read every time. Use a cache where relevant.
-(define (load-xiden-rcfile)
-  (let ([path (get-xiden-settings-path)])
-    (if (file-exists? path)
-        (parameterize ([sandbox-path-permissions
-                        (append (map (λ (p) `(exists ,p)) (filesystem-root-list/cached))
-                                (sandbox-path-permissions))])
-          (make-module-evaluator #:language 'xiden/rcfile path))
-        void)))
-
-
-(define (call-with-rcfile proc)
-  (parameterize ([current-xiden-rcfile-cache (load-xiden-rcfile)])
-    (proc)))
 
 
 (define (dump-xiden-settings)
@@ -109,7 +87,6 @@
 (define-xiden-setting XIDEN_INSTALL_SOURCES (listof (list/c string? string? string?)) null)
 (define-xiden-setting XIDEN_MEMORY_LIMIT_MB (>=/c 0) 200)
 (define-xiden-setting XIDEN_MESSAGE_DIGEST_ALGORITHM md-algorithm/c 'sha384)
-(define-xiden-setting XIDEN_PLUGIN_MODULE (or/c #f path-string?) #f)
 (define-xiden-setting XIDEN_READER_FRIENDLY_OUTPUT boolean? #f)
 (define-xiden-setting XIDEN_SIGNER (list/c (or/c #f string?) (or/c #f path-string?) (or/c #f path-string?)) '(#f #f #f))
 (define-xiden-setting XIDEN_SUBPROCESS_TIMEOUT_S (>=/c 0) (* 30 60))
@@ -127,25 +104,14 @@
 
 (module+ test
   (require rackunit
-           (submod "file.rkt" test))
-
+           (submod "file.rkt" test)
+           (submod "plugin.rkt" test))
 
   (test-not-exn "Group all runtime configuration in a hash table"
                 (λ ()
                   (invariant-assertion
                    (and/c immutable? (hash/c symbol? setting?))
                    XIDEN_SETTINGS)))
-
-  (test-case "Define runtime configuration file path in terms of workspace"
-    (define rcfile-/a
-      (parameterize ([workspace-directory "/a"])
-        (get-xiden-settings-path)))
-    (define rcfile-/b
-      (parameterize ([workspace-directory "/b"])
-        (get-xiden-settings-path)))
-    (check-pred complete-path? rcfile-/a)
-    (check-pred complete-path? rcfile-/b)
-    (check-not-equal? rcfile-/a rcfile-/b))
 
 
   (test-case "Dump current runtime configuration to a hash table"
@@ -156,38 +122,37 @@
         (dump-xiden-settings))))
 
     (define dump (dump-xiden-settings))
-    (check-equal? (hash-ref dump 'XIDEN_PLUGIN_MODULE)
-                  (XIDEN_PLUGIN_MODULE))
+    (check-equal? (hash-ref dump 'XIDEN_GENERATED_INPUT_NAME)
+                  (XIDEN_GENERATED_INPUT_NAME))
 
-    (XIDEN_PLUGIN_MODULE
+    (XIDEN_GENERATED_INPUT_NAME
      "foo"
      (λ ()
-       (check-not-equal? (hash-ref dump 'XIDEN_PLUGIN_MODULE)
-                         (XIDEN_PLUGIN_MODULE))
-       (check-equal? (hash-ref (dump-xiden-settings) 'XIDEN_PLUGIN_MODULE)
-                     (XIDEN_PLUGIN_MODULE)))))
+       (check-not-equal? (hash-ref dump 'XIDEN_GENERATED_INPUT_NAME)
+                         (XIDEN_GENERATED_INPUT_NAME))
+       (check-equal? (hash-ref (dump-xiden-settings) 'XIDEN_GENERATED_INPUT_NAME)
+                     (XIDEN_GENERATED_INPUT_NAME)))))
 
   (test-workspace "Find fallback values from several sources"
-    (test-false "First use a hard-coded value"
-                (XIDEN_PLUGIN_MODULE))
+    (test-equal? "First use a hard-coded value"
+                 (XIDEN_GENERATED_INPUT_NAME)
+                 DEFAULT_STRING)
 
     (test-case "Override hard-coded value with rcfile"
-      (make-directory* (path-only (get-xiden-settings-path)))
-      (write-to-file
-       `(module rc xiden/rcfile (define XIDEN_PLUGIN_MODULE "foo"))
-       (get-xiden-settings-path))
-      (call-with-rcfile
-       (λ () (check-equal? (XIDEN_PLUGIN_MODULE) "foo"))))
-
-    (test-case "Override rcfile value with envvar value"
-      (dynamic-wind (λ () (putenv "XIDEN_PLUGIN_MODULE" "\"bar\""))
-                    (λ () (check-equal? (XIDEN_PLUGIN_MODULE) "bar"))
-                    (λ ()
-                      (putenv "XIDEN_PLUGIN_MODULE" "")
-                      (call-with-rcfile
-                        (λ () (test-equal? "Do not use empty envvar strings as a source for settings"
-                                           (XIDEN_PLUGIN_MODULE)
-                                           "foo")))))))
+      (call-with-plugin
+       `(module rc racket/base
+          (provide (all-defined-out))
+          (define XIDEN_GENERATED_INPUT_NAME "foo"))
+       (λ ()
+         (check-equal? (XIDEN_GENERATED_INPUT_NAME) "foo")
+         (test-case "Override rcfile value with envvar value"
+           (dynamic-wind (λ () (putenv "XIDEN_GENERATED_INPUT_NAME" "\"bar\""))
+                         (λ () (check-equal? (XIDEN_GENERATED_INPUT_NAME) "bar"))
+                         (λ ()
+                           (putenv "XIDEN_GENERATED_INPUT_NAME" "")
+                           (test-equal? "Do not use empty envvar strings as a source for settings"
+                                        (XIDEN_GENERATED_INPUT_NAME)
+                                        "foo"))))))))
 
   (test-workspace "Lazily validate fallback values"
     (test-exn "Reject invalid values from ennvar"
@@ -195,7 +160,7 @@
               (λ ()
                 (dynamic-wind void
                               (λ ()
-                                (putenv "XIDEN_PLUGIN_MODULE" "123")
-                                (XIDEN_PLUGIN_MODULE))
+                                (putenv "XIDEN_GENERATED_INPUT_NAME" "123")
+                                (XIDEN_GENERATED_INPUT_NAME))
                               (λ ()
-                                (putenv "XIDEN_PLUGIN_MODULE" "")))))))
+                                (putenv "XIDEN_GENERATED_INPUT_NAME" "")))))))